+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/ErrorManagement/Persistence/ElasticsearchBusinessErrorRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\Persistence\Elasticsearch\Repository;
+namespace Akeneo\Connectivity\Connection\Infrastructure\ErrorManagement\Persistence;
 
 use Akeneo\Connectivity\Connection\Domain\ErrorManagement\Model\Write\BusinessError;
 use Akeneo\Connectivity\Connection\Domain\ErrorManagement\Persistence\Repository\BusinessErrorRepositoryInterface;
