+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/queries.yml
@@ -9,21 +9,6 @@ services:
         arguments:
             - '@database_connection'
 
-    akeneo_connectivity_connection.persistence.query.select_last_connection_business_errors:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Elasticsearch\Query\ElasticsearchSelectLastConnectionBusinessErrorsQuery
-        arguments:
-            - '@akeneo_connectivity.client.connection_error'
-
-    akeneo_connectivity_connection.persistence.query.select_all_auditable_connection_code:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Query\SelectAllAuditableConnectionCodeQuery
-        arguments:
-            - '@database_connection'
-
-    akeneo_connectivity_connection.persistence.query.purge_connection_errors:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Elasticsearch\Query\PurgeConnectionErrorsQuery
-        arguments:
-            - '@akeneo_connectivity.client.connection_error'
-
     Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Query\IsConnectionsNumberLimitReachedQuery:
         arguments:
             - '@database_connection'
