+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/repositories.yml
@@ -3,13 +3,3 @@ services:
         class: 'Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Repository\DbalConnectionRepository'
         arguments:
             - '@database_connection'
-
-    akeneo_connectivity.connection.persistence.repository.error_count:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Repository\DbalErrorCountRepository
-        arguments:
-            - '@database_connection'
-
-    akeneo_connectivity.connection.persistence.repository.business_error:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Elasticsearch\Repository\ElasticsearchBusinessErrorRepository
-        arguments:
-            - '@akeneo_connectivity.client.connection_error'
