+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/services.yml
@@ -69,7 +69,7 @@ services:
         class: Akeneo\Connectivity\Connection\Application\ErrorManagement\Service\CollectApiError
         arguments:
             - '@akeneo_connectivity.connection.service.connection_context'
-            - '@akeneo_connectivity.connection.persistence.repository.business_error'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\ErrorManagement\Persistence\ElasticsearchBusinessErrorRepository'
             - '@akeneo_connectivity.connection.application.error_management.handler.update_connection_error_count'
             - '@akeneo_connectivity.connection.exception_serializer'
