+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/ErrorManagement/Persistence/DbalErrorCountRepositoryIntegration.php
@@ -1,13 +1,13 @@
 <?php
+
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Tests\Integration\Audit\Persistence;
+namespace Akeneo\Connectivity\Connection\Tests\Integration\ErrorManagement\Persistence;
 
 use Akeneo\Connectivity\Connection\Domain\ErrorManagement\ErrorTypes;
 use Akeneo\Connectivity\Connection\Domain\ErrorManagement\Model\Write\HourlyErrorCount;
 use Akeneo\Connectivity\Connection\Domain\ValueObject\HourlyInterval;
-use Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Repository\DbalErrorCountRepository;
-use Akeneo\Connectivity\Connection\Tests\CatalogBuilder\ConnectionLoader;
+use Akeneo\Connectivity\Connection\Infrastructure\ErrorManagement\Persistence\DbalErrorCountRepository;
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
 use Doctrine\DBAL\Connection as DbalConnection;
@@ -15,14 +15,8 @@ use PHPUnit\Framework\Assert;
 
 class DbalErrorCountRepositoryIntegration extends TestCase
 {
-    /** @var DbalConnection */
-    private $dbalConnection;
-
-    /** @var DbalErrorCountRepository */
-    private $repository;
-
-    /** @var ConnectionLoader */
-    private $connectionLoader;
+    private DbalConnection $dbalConnection;
+    private DbalErrorCountRepository $repository;
 
     public function test_it_inserts_an_error_count()
     {
@@ -93,8 +87,7 @@ SQL;
         parent::setUp();
 
         $this->dbalConnection = $this->get('database_connection');
-        $this->repository = $this->get('akeneo_connectivity.connection.persistence.repository.error_count');
-        $this->connectionLoader = $this->get('akeneo_connectivity.connection.fixtures.connection_loader');
+        $this->repository = $this->get(DbalErrorCountRepository::class);
     }
 
     protected function getConfiguration(): Configuration
